refactor: useFocusControlをMultiCombobox内に隠蔽する
